requirements.txt
alembic==1.3.1
astroid==2.3.3
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
paramiko==2.6.0
passlib==1.7.1
psutil==5.5.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.19
pylint==2.4.4
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2018.9
simplejson==3.16.0
six==1.13.0
speaklater==1.3
SQLAlchemy==1.3.11
sqlparse==0.2.4
sshtunnel==0.1.5
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
